+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add PostgreSQLAdapter#set_replica_identity
 - Add methods for discovering and managing extensions
 - Add method to temporarily add a schema to the search path
+- Add PostgreSQLAdapter#vacuum
 
 ## [0.1.1] - 2021-06-07